Move some U256 conversions into own crate
