+++ crates/orderbook/Cargo.toml
@@ -32,6 +32,7 @@ hex-literal = "0.3"
 maplit = "1.0"
 model = { path = "../model" }
 num = "0.4"
+number_conversions = { path = "../number_conversions" }
 primitive-types = { version = "0.10", features = ["fp-conversion"] }
 prometheus = "0.13"
 prometheus-metric-storage = { git = "https://github.com/cowprotocol/prometheus-metric-storage" , tag = "v0.4.0" }
