+++ crates/orderbook/src/conversions.rs
@@ -1,29 +1,16 @@
 use anyhow::{anyhow, Result};
 use bigdecimal::num_bigint::ToBigInt;
-use num::bigint::{BigInt, BigUint, Sign};
+use num::bigint::{BigInt, BigUint};
 use primitive_types::{H160, H256, U256};
 use sqlx::types::BigDecimal;
 use std::convert::TryInto;
+
 // TODO: It would be nice to avoid copying the underlying BigInt when converting BigDecimal to
 // anything else but the simple big_decimal.to_bigint makes a copy internally.
 
-pub fn u256_to_big_uint(input: &U256) -> BigUint {
-    let mut bytes = [0; 32];
-    input.to_big_endian(&mut bytes);
-    BigUint::from_bytes_be(&bytes)
-}
-
 pub fn u256_to_big_decimal(u256: &U256) -> BigDecimal {
-    let big_uint = u256_to_big_uint(u256);
+    let big_uint = number_conversions::u256_to_big_uint(u256);
     BigDecimal::from(BigInt::from(big_uint))
-}
-
-pub fn bigint_to_u256(input: &BigInt) -> Option<U256> {
-    let (sign, bytes) = input.to_bytes_be();
-    if sign == Sign::Minus || bytes.len() > 32 {
-        return None;
-    }
-    Some(U256::from_big_endian(&bytes))
 }
 
 pub fn big_decimal_to_big_uint(big_decimal: &BigDecimal) -> Option<BigUint> {
@@ -35,7 +22,7 @@ pub fn big_decimal_to_u256(big_decimal: &BigDecimal) -> Option<U256> {
         return None;
     }
     let big_int = big_decimal.to_bigint()?;
-    bigint_to_u256(&big_int)
+    number_conversions::big_int_to_u256(&big_int).ok()
 }
 
 pub fn h160_from_vec(vec: Vec<u8>) -> Result<H160> {
@@ -59,19 +46,6 @@ mod tests {
     use std::str::FromStr;
 
     #[test]
-    fn u256_to_big_uint_() {
-        assert_eq!(u256_to_big_uint(&U256::zero()), BigUint::zero());
-        assert_eq!(u256_to_big_uint(&U256::one()), BigUint::one());
-        assert_eq!(
-            u256_to_big_uint(&U256::MAX),
-            BigUint::from_str(
-                "115792089237316195423570985008687907853269984665640564039457584007913129639935"
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
     fn u256_to_big_decimal_() {
         assert_eq!(u256_to_big_decimal(&U256::zero()), BigDecimal::zero());
         assert_eq!(u256_to_big_decimal(&U256::one()), BigDecimal::one());
@@ -82,21 +56,6 @@ mod tests {
             )
             .unwrap()
         );
-    }
-
-    #[test]
-    fn bigint_to_u256_() {
-        assert_eq!(bigint_to_u256(&BigInt::zero()), Some(U256::zero()));
-        assert_eq!(bigint_to_u256(&BigInt::one()), Some(U256::one()));
-
-        let max_u256_as_bigint = BigInt::from_str(
-            "115792089237316195423570985008687907853269984665640564039457584007913129639935",
-        )
-        .unwrap();
-        assert_eq!(bigint_to_u256(&max_u256_as_bigint), Some(U256::MAX));
-
-        assert!(bigint_to_u256(&(max_u256_as_bigint + BigInt::one())).is_none());
-        assert!(bigint_to_u256(&BigInt::from(-1)).is_none());
     }
 
     #[test]
