+++ crates/shared/Cargo.toml
@@ -32,6 +32,7 @@ maplit = "1.0"
 mockall = "0.11"
 model = { path = "../model" }
 num = "0.4"
+number_conversions = { path = "../number_conversions" }
 once_cell = "1.9.0"
 primitive-types = "0.10"
 prometheus = "0.13"
