+++ crates/solver/Cargo.toml
@@ -18,6 +18,7 @@ path = "src/main.rs"
 anyhow = "1.0"
 async-trait = "0.1"
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
+clap = { version = "3.1", features = ["derive", "env"] }
 contracts = { path = "../contracts" }
 derivative = "2.2"
 derive_more = "0.99"
@@ -33,6 +34,7 @@ lazy_static = "1.4"
 maplit = "1.0"
 model = { path = "../model" }
 num = "0.4"
+number_conversions = { path = "../number_conversions" }
 primitive-types = { version = "0.10", features = ["fp-conversion"] }
 prometheus = "0.13"
 prometheus-metric-storage = { git = "https://github.com/cowprotocol/prometheus-metric-storage" , tag = "v0.4.0" }
@@ -42,7 +44,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_with = { version = "1.11", default-features = false }
 shared = { path = "../shared" }
-clap = { version = "3.1", features = ["derive", "env"] }
 strum = { version = "0.24", features = ["derive"] }
 thiserror = "1.0"
 tokio = { version = "1.15", features = ["macros", "rt-multi-thread", "time", "test-util"] }
